test: split child agent white-box suites
