+++ crates/agent-engine/src/runtime/child_agents.rs
@@ -1421,2649 +1421,4 @@ fn truncate_chars(text: &str, limit: usize) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::llm::{GenerationRequest, GenerationResponse, StreamChunk, TokenUsage};
-    use crate::runtime::{
-        ApprovedMountGrant, ApprovedMountGrantAccess, MountGrantApplicator,
-        MountGrantApplicatorFactory, NamespaceRuntimeEnvironment, RuntimeConfig,
-    };
-    use crate::skills::SkillHostCapabilities;
-    use crate::tools::Tool;
-    use crate::tools::ToolRegistry;
-    use alan_ap::{Fid, FileServer, InProcessTransport, OpenMode};
-    use alan_kernel::{
-        Access as KernelAccess, Credentials as KernelCredentials, Namespace as KernelNamespace,
-        ProcFs as KernelProcFs,
-    };
-    use alan_llm::LlmProvider;
-    use serde_json::json;
-    use std::sync::{Arc, Mutex};
-    use tempfile::TempDir;
-
-    fn test_startup_metadata(
-        process_path: impl Into<String>,
-        rollout_path: Option<PathBuf>,
-        durable: bool,
-    ) -> RuntimeStartupMetadata {
-        RuntimeStartupMetadata {
-            process_path: process_path.into(),
-            agent_path: "/agent/test".to_string(),
-            rollout_id: None,
-            rollout_path,
-            durability: super::super::engine::AgentMachineDurabilityState {
-                durable,
-                required: false,
-            },
-            execution_backend: crate::tools::active_backend_name().to_string(),
-            request_controls: crate::ResolvedRequestControls::default(),
-            warnings: Vec::new(),
-        }
-    }
-
-    fn namespace_environment_for_parent_test_with_route(
-        routefs: Arc<alan_routefs::RouteFs>,
-    ) -> NamespaceRuntimeEnvironment {
-        namespace_environment_for_parent_test_with_services(
-            routefs,
-            Arc::new(alan_llmfs::LlmFs::new()),
-        )
-    }
-
-    fn namespace_environment_for_parent_test_with_services(
-        routefs: Arc<alan_routefs::RouteFs>,
-        llmfs: Arc<alan_llmfs::LlmFs>,
-    ) -> NamespaceRuntimeEnvironment {
-        namespace_environment_for_parent_test_with_connection(routefs, llmfs, "default")
-    }
-
-    fn namespace_environment_for_parent_test_with_connection(
-        routefs: Arc<alan_routefs::RouteFs>,
-        llmfs: Arc<alan_llmfs::LlmFs>,
-        connection: &str,
-    ) -> NamespaceRuntimeEnvironment {
-        let mut mounts = KernelNamespace::new();
-        for name in ["alpha", "beta"] {
-            let manifest =
-                crate::runtime::ToolPackageManifest::from_tool(&NamedTestTool::new(name), 30)
-                    .unwrap();
-            mounts.mount(
-                &format!("/bin/{name}"),
-                memfs_transport(),
-                KernelAccess::ReadOnly,
-            );
-            mounts.mount(
-                &format!("/lib/exec/{name}"),
-                InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::with_read_only_file(
-                    "manifest",
-                    serde_json::to_vec(&manifest).unwrap(),
-                ))),
-                KernelAccess::ReadOnly,
-            );
-        }
-        let root = InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(mounts)));
-        let procfs = KernelProcFs::new();
-        let tool_runner = crate::tools::ToolRegistry::new().process_runner();
-        let assembler = TestChildProcessAssembler {
-            procfs,
-            tool_runner,
-            srv: memfs_transport(),
-            route: InProcessTransport::new(routefs),
-            llm: InProcessTransport::new(llmfs),
-            parent: None,
-        };
-        crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", connection)
-            .with_child_process_assembler(Arc::new(assembler))
-    }
-
-    #[derive(Clone)]
-    struct TestChildProcessAssembler {
-        procfs: KernelProcFs,
-        tool_runner: crate::tools::ToolProcessRunner,
-        srv: InProcessTransport,
-        route: InProcessTransport,
-        llm: InProcessTransport,
-        parent: Option<TestParentProcessContext>,
-    }
-
-    impl std::fmt::Debug for TestChildProcessAssembler {
-        fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            formatter
-                .debug_struct("TestChildProcessAssembler")
-                .finish_non_exhaustive()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl super::super::ChildAgentProcessAssembler for TestChildProcessAssembler {
-        async fn assemble(
-            &self,
-            request: super::super::ChildAgentProcessAssemblyRequest,
-        ) -> Result<super::super::AssembledChildAgentProcess> {
-            let super::super::ChildAgentProcessAssemblyRequest {
-                plan,
-                scratch_dir,
-                executable,
-                llm_override,
-            } = request;
-            let mut handles = ChildNamespaceLaunchHandles::new(
-                Arc::new(alan_agentfs::AgentFs::new()),
-                llm_override.unwrap_or_else(|| self.llm.clone()),
-                self.srv.clone(),
-                self.route.clone(),
-            );
-            for manifest in &plan.tool_packages {
-                let name = &manifest.name;
-                handles = handles.with_tool_package(
-                    format!("/bin/{name}"),
-                    memfs_transport(),
-                    format!("/lib/exec/{name}"),
-                    InProcessTransport::new(Arc::new(
-                        alan_ap::reference::MemFs::with_read_only_file(
-                            "manifest",
-                            serde_json::to_vec(manifest)?,
-                        ),
-                    )),
-                );
-            }
-            let runtime_procfs = self
-                .procfs
-                .clone()
-                .with_runner(Arc::new(self.tool_runner.clone()));
-            let binding = plan.runtime_execution_binding(scratch_dir)?;
-            let mut launch = spawn_child_namespace_runtime_environment(
-                &self.procfs,
-                &runtime_procfs,
-                &plan,
-                handles,
-                self.parent.clone(),
-                self.tool_runner.clone(),
-                binding,
-                None,
-                &executable,
-            )
-            .await?;
-            let observation_environment = child_observation_environment(
-                &runtime_procfs,
-                launch.agent_root.clone(),
-                &launch.pid,
-                &plan,
-            )
-            .await?;
-            let child_assembler = Self {
-                procfs: self.procfs.clone(),
-                tool_runner: self.tool_runner.clone(),
-                srv: self.srv.clone(),
-                route: self.route.clone(),
-                llm: self.llm.clone(),
-                parent: Some(TestParentProcessContext {
-                    agent_root: launch.agent_root.clone(),
-                    pid: alan_kernel::Pid(launch.pid.parse()?),
-                }),
-            };
-            launch.environment = launch
-                .environment
-                .with_child_process_assembler(Arc::new(child_assembler));
-            Ok(super::super::AssembledChildAgentProcess {
-                pid: launch.pid,
-                environment: launch.environment,
-                observation_environment,
-                lifecycle: launch.lifecycle,
-            })
-        }
-    }
-
-    #[derive(Clone, Default)]
-    struct RecordedRequests(Arc<Mutex<Vec<GenerationRequest>>>);
-
-    #[derive(Clone)]
-    struct RecordingProvider {
-        requests: RecordedRequests,
-        response: GenerationResponse,
-        delay: Option<Duration>,
-    }
-
-    impl RecordingProvider {
-        fn new(requests: RecordedRequests, response: GenerationResponse) -> Self {
-            Self {
-                requests,
-                response,
-                delay: None,
-            }
-        }
-
-        fn with_delay(mut self, delay: Duration) -> Self {
-            self.delay = Some(delay);
-            self
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl LlmProvider for RecordingProvider {
-        async fn generate(
-            &mut self,
-            request: GenerationRequest,
-        ) -> anyhow::Result<GenerationResponse> {
-            self.requests.0.lock().unwrap().push(request);
-            if let Some(delay) = self.delay {
-                tokio::time::sleep(delay).await;
-            }
-            Ok(self.response.clone())
-        }
-
-        async fn chat(&mut self, _system: Option<&str>, user: &str) -> anyhow::Result<String> {
-            Ok(format!("chat: {user}"))
-        }
-
-        async fn generate_stream(
-            &mut self,
-            request: GenerationRequest,
-        ) -> anyhow::Result<tokio::sync::mpsc::Receiver<StreamChunk>> {
-            self.requests.0.lock().unwrap().push(request);
-            if let Some(delay) = self.delay {
-                tokio::time::sleep(delay).await;
-            }
-            let (tx, rx) = tokio::sync::mpsc::channel(1);
-            let _ = tx
-                .send(StreamChunk {
-                    text: Some(self.response.content.clone()),
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: None,
-                    usage: Some(TokenUsage {
-                        prompt_tokens: 1,
-                        cached_prompt_tokens: None,
-                        completion_tokens: 1,
-                        total_tokens: 2,
-                        reasoning_tokens: None,
-                    }),
-                    provider_response_id: None,
-                    provider_response_status: None,
-                    sequence_number: None,
-                    tool_call_delta: None,
-                    is_finished: true,
-                    finish_reason: Some("stop".to_string()),
-                })
-                .await;
-            Ok(rx)
-        }
-
-        fn provider_name(&self) -> &'static str {
-            "openai_responses"
-        }
-    }
-
-    struct NamedTestTool {
-        name: String,
-    }
-
-    impl NamedTestTool {
-        fn new(name: &str) -> Self {
-            Self {
-                name: name.to_string(),
-            }
-        }
-    }
-
-    impl Tool for NamedTestTool {
-        fn name(&self) -> &str {
-            &self.name
-        }
-
-        fn description(&self) -> &str {
-            "test tool"
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            json!({
-                "type": "object",
-                "properties": {}
-            })
-        }
-
-        fn execute(
-            &self,
-            _arguments: serde_json::Value,
-            _ctx: &crate::tools::ToolContext,
-        ) -> crate::tools::ToolResult {
-            Box::pin(async { Ok(json!({"ok": true})) })
-        }
-    }
-
-    #[derive(Debug, Default)]
-    struct RecordingMountGrantApplicatorFactory {
-        created: Arc<Mutex<usize>>,
-        applied: Arc<Mutex<Vec<ApprovedMountGrant>>>,
-    }
-
-    impl RecordingMountGrantApplicatorFactory {
-        fn created_count(&self) -> usize {
-            *self
-                .created
-                .lock()
-                .expect("created count lock should not be poisoned")
-        }
-
-        fn applied_grants(&self) -> Vec<ApprovedMountGrant> {
-            self.applied
-                .lock()
-                .expect("applied grants lock should not be poisoned")
-                .clone()
-        }
-    }
-
-    impl MountGrantApplicatorFactory for RecordingMountGrantApplicatorFactory {
-        fn create(
-            &self,
-            _pid: alan_kernel::Pid,
-            live_namespace: alan_kernel::LiveNamespace,
-            _inherited_mount_paths: &[String],
-        ) -> Arc<dyn MountGrantApplicator> {
-            *self
-                .created
-                .lock()
-                .expect("created count lock should not be poisoned") += 1;
-            Arc::new(RecordingMountGrantApplicator {
-                live_namespace,
-                applied: self.applied.clone(),
-            })
-        }
-    }
-
-    #[derive(Debug)]
-    struct RecordingMountGrantApplicator {
-        live_namespace: alan_kernel::LiveNamespace,
-        applied: Arc<Mutex<Vec<ApprovedMountGrant>>>,
-    }
-
-    impl MountGrantApplicator for RecordingMountGrantApplicator {
-        fn apply_mount_grant(&self, grant: &ApprovedMountGrant) -> anyhow::Result<KernelNamespace> {
-            self.applied
-                .lock()
-                .expect("applied grants lock should not be poisoned")
-                .push(grant.clone());
-            let access = match grant.access {
-                ApprovedMountGrantAccess::ReadOnly => KernelAccess::ReadOnly,
-                ApprovedMountGrantAccess::ReadWrite => KernelAccess::ReadWrite,
-            };
-            self.live_namespace
-                .mount(&grant.namespace_path, memfs_transport(), access);
-            Ok(self.live_namespace.snapshot())
-        }
-    }
-
-    struct MarkerTool {
-        name: String,
-        marker: String,
-    }
-
-    impl MarkerTool {
-        fn new(name: &str, marker: &str) -> Self {
-            Self {
-                name: name.to_string(),
-                marker: marker.to_string(),
-            }
-        }
-    }
-
-    impl Tool for MarkerTool {
-        fn name(&self) -> &str {
-            &self.name
-        }
-
-        fn description(&self) -> &str {
-            "marker test tool"
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            json!({
-                "type": "object",
-                "properties": {}
-            })
-        }
-
-        fn execute(
-            &self,
-            _arguments: serde_json::Value,
-            _ctx: &crate::tools::ToolContext,
-        ) -> crate::tools::ToolResult {
-            let marker = self.marker.clone();
-            Box::pin(async move { Ok(json!({ "marker": marker })) })
-        }
-    }
-
-    fn make_parent_state(
-        temp: &TempDir,
-        requests: RecordedRequests,
-        response: GenerationResponse,
-    ) -> RuntimeLoopState {
-        make_parent_state_with_capability_view(
-            temp,
-            requests,
-            response,
-            crate::skills::ResolvedCapabilityView::default(),
-        )
-    }
-
-    fn make_parent_state_with_capability_view(
-        temp: &TempDir,
-        requests: RecordedRequests,
-        response: GenerationResponse,
-        capability_view: crate::skills::ResolvedCapabilityView,
-    ) -> RuntimeLoopState {
-        let source_root = temp.path().join("source");
-        let definition_root = temp.path().join("definition");
-        let store_root = temp.path().join("system-store");
-        std::fs::create_dir_all(&source_root).unwrap();
-        std::fs::create_dir_all(definition_root.join("persona")).unwrap();
-        std::fs::create_dir_all(definition_root.join("skills")).unwrap();
-        std::fs::write(
-            definition_root.join("agent.toml"),
-            "tool_repeat_limit = 4\n",
-        )
-        .unwrap();
-
-        let store_bindings = crate::AgentRuntimeStoreBindings {
-            rollouts: store_root.join("rollouts"),
-            checkpoints: store_root.join("checkpoints"),
-            cache: store_root.join("cache"),
-            tmp: store_root.join("tmp"),
-            metadata: store_root.join("metadata"),
-        };
-        for path in [
-            &store_bindings.rollouts,
-            &store_bindings.checkpoints,
-            &store_bindings.cache,
-            &store_bindings.tmp,
-            &store_bindings.metadata,
-        ] {
-            std::fs::create_dir_all(path).unwrap();
-        }
-        let memory_store = store_root.join("memory");
-        std::fs::create_dir_all(&memory_store).unwrap();
-
-        let mut launch_namespace = KernelNamespace::new();
-        launch_namespace.mount("/mnt/source", memfs_transport(), KernelAccess::ReadWrite);
-        launch_namespace.mount(
-            "/agent-definition",
-            memfs_transport(),
-            KernelAccess::ReadOnly,
-        );
-        launch_namespace.mount("/memory", memfs_transport(), KernelAccess::ReadWrite);
-        let launch_context = crate::ProcessLaunchContext::new(
-            launch_namespace,
-            KernelCredentials::user("parent-agent"),
-            "/mnt/source",
-        )
-        .unwrap()
-        .with_host_mount(
-            crate::HostMountGrant::new("/mnt/source", &source_root, KernelAccess::ReadWrite)
-                .unwrap(),
-        )
-        .with_host_mount(
-            crate::HostMountGrant::new(
-                "/agent-definition",
-                &definition_root,
-                KernelAccess::ReadOnly,
-            )
-            .unwrap(),
-        )
-        .with_descriptor(
-            crate::AGENT_DEFINITION_DESCRIPTOR,
-            crate::ProcessDescriptor::new("/agent-definition").unwrap(),
-        )
-        .with_descriptor(
-            crate::MEMORY_STORE_DESCRIPTOR,
-            crate::ProcessDescriptor::new("/memory").unwrap(),
-        );
-
-        let mut core_config = crate::Config::default();
-        core_config.memory.store_dir = Some(memory_store.clone());
-        core_config.openai_responses_model = "gpt-5.4".to_string();
-        let mut machine = crate::AgentMachine::new();
-        machine.add_user_message("Parent user asks for review");
-        machine.add_assistant_message("Parent assistant explains the approach", None);
-        machine.add_tool_message("tool_call_1", "alpha", json!({"summary": "tool output"}));
-
-        let mut turn_state = super::super::TurnState::default();
-        turn_state.set_plan_snapshot(
-            Some("Finish the delegated check".to_string()),
-            vec![alan_agent_protocol::PlanItem {
-                id: "plan-1".to_string(),
-                content: "Inspect the changed files".to_string(),
-                status: alan_agent_protocol::PlanItemStatus::InProgress,
-            }],
-        );
-
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        llmfs.register_connection(
-            "default",
-            Box::new(RecordingProvider::new(requests, response)),
-        );
-
-        RuntimeLoopState {
-            machine,
-            current_submission_id: None,
-            environment: namespace_environment_for_parent_test_with_services(
-                Arc::new(alan_routefs::RouteFs::new()),
-                llmfs,
-            )
-            .with_launch_context(launch_context),
-            core_config,
-            runtime_config: RuntimeConfig {
-                store_bindings: Some(store_bindings),
-                memory_store_backing: Some(memory_store),
-                ..RuntimeConfig::default()
-            },
-            definition_persona_dirs: Vec::new(),
-            prompt_cache:
-                super::super::prompt_cache::PromptAssemblyCache::with_fixed_capability_view(
-                    capability_view,
-                    Vec::new(),
-                    SkillHostCapabilities::with_tools(["alpha", "beta"]),
-                ),
-            turn_state,
-        }
-    }
-
-    fn parent_test_tools(config: &crate::Config) -> ToolRegistry {
-        let mut tools = ToolRegistry::with_config(Arc::new(config.clone()));
-        tools.register(NamedTestTool::new("alpha"));
-        tools.register(NamedTestTool::new("beta"));
-        tools
-    }
-
-    fn inherited_launch_context(parent: &RuntimeLoopState) -> crate::ProcessLaunchContext {
-        parent
-            .namespace_environment()
-            .launch_context()
-            .expect("test parent has a Process Launch Context")
-            .child()
-    }
-
-    fn launch_spec(_definition_root: PathBuf) -> SpawnSpec {
-        SpawnSpec {
-            target: SpawnTarget::DefinitionDescriptor {
-                descriptor: crate::AGENT_DEFINITION_DESCRIPTOR.to_string(),
-            },
-            launch: alan_agent_protocol::SpawnLaunchInputs {
-                task: "Review the repository changes".to_string(),
-                cwd: None,
-                timeout_secs: Some(30),
-                output_dir: None,
-            },
-            handles: vec![SpawnHandle::HostMounts],
-            runtime_overrides: alan_agent_protocol::SpawnRuntimeOverrides::default(),
-            delegated: None,
-        }
-    }
-
-    fn capability_plan(host_mount: Option<PathBuf>, tools: &[&str]) -> ChildNamespaceAssemblyPlan {
-        let mut launch_context = crate::ProcessLaunchContext::root();
-        let cwd = host_mount.as_ref().map(|_| PathBuf::from("/mnt/source"));
-        if let Some(host_mount) = host_mount {
-            launch_context.namespace.mount(
-                "/mnt/source",
-                memfs_transport(),
-                KernelAccess::ReadWrite,
-            );
-            launch_context.host_mounts.push(
-                crate::HostMountGrant::new("/mnt/source", host_mount, KernelAccess::ReadWrite)
-                    .unwrap(),
-            );
-            launch_context.cwd = "/mnt/source".to_string();
-        }
-        ChildNamespaceAssemblyPlan {
-            agent_mount: "/agent".to_string(),
-            llm_mount: "/mnt/llm".to_string(),
-            llm_connection_name: "default".to_string(),
-            srv_mount: "/srv".to_string(),
-            route_mount: alan_routefs::MOUNT_PATH.to_string(),
-            bin_tool_mounts: tools.iter().map(|tool| format!("/bin/{tool}")).collect(),
-            tool_packages: Vec::new(),
-            cwd,
-            launch_context,
-        }
-    }
-
-    #[test]
-    fn inherited_mount_without_host_backed_cwd_uses_authorized_native_cwd() {
-        let source = TempDir::new().unwrap();
-        let scratch = TempDir::new().unwrap();
-        let mut plan = capability_plan(Some(source.path().to_path_buf()), &["read_file"]);
-        plan.cwd = None;
-        plan.launch_context.cwd = "/".to_string();
-
-        let binding = plan
-            .runtime_execution_binding(Some(scratch.path().to_path_buf()))
-            .unwrap()
-            .expect("an inherited Host Mount should create a child Tool binding");
-
-        assert_eq!(binding.cwd, dunce::canonicalize(source.path()).unwrap());
-        assert_eq!(binding.namespace_cwd, PathBuf::from("/mnt/source"));
-        assert_eq!(binding.host_mounts.len(), 1);
-        assert_eq!(binding.host_mounts[0].namespace_path, "/mnt/source");
-        let sandbox = binding.sandbox_spec.unwrap();
-        assert!(
-            !sandbox
-                .readable_roots
-                .iter()
-                .any(|root| root == &dunce::canonicalize(scratch.path()).unwrap())
-        );
-    }
-
-    #[test]
-    fn package_projection_does_not_create_host_tool_authority() {
-        let mut plan = capability_plan(None, &["read_file"]);
-        plan.launch_context.package_references.push(
-            crate::ProcessPackageReference::new(
-                "example",
-                "a".repeat(64),
-                crate::ProcessPackageKind::Installed,
-                "/lib/pkg/example",
-                Vec::new(),
-                memfs_transport(),
-            )
-            .unwrap(),
-        );
-
-        assert!(plan.runtime_execution_binding(None).unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn delegated_spawn_boundary_passes_satisfied_task_unchanged() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("unused"),
-        );
-        let host_mount = PathBuf::from("/tmp/repo");
-        let mut spec = launch_spec(temp.path().join("agent"));
-        spec.launch.task = "Inspect local files".to_string();
-        spec.delegated = Some(alan_agent_protocol::DelegatedSpawnContext {
-            requirements: vec![
-                alan_agent_protocol::DelegatedCapabilityRequirement::MountRead {
-                    path: Some(PathBuf::from("/mnt/source")),
-                },
-                alan_agent_protocol::DelegatedCapabilityRequirement::LlmConnection,
-            ],
-        });
-
-        let decision = evaluate_delegated_launch_capabilities(
-            &parent,
-            &mut spec,
-            &capability_plan(Some(host_mount), &["read_file"]),
-        )
-        .await
-        .unwrap()
-        .unwrap();
-
-        assert_eq!(
-            decision.recovery,
-            alan_agent_protocol::DelegatedCapabilityRecovery::Satisfied
-        );
-        assert_eq!(spec.launch.task, "Inspect local files");
-    }
-
-    #[tokio::test]
-    async fn delegated_spawn_boundary_rewrites_narrowed_task_explicitly() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("unused"),
-        );
-        let host_mount = PathBuf::from("/tmp/repo");
-        let mut spec = launch_spec(temp.path().join("agent"));
-        spec.launch.task = "Review GitHub issue against local code".to_string();
-        spec.delegated = Some(alan_agent_protocol::DelegatedSpawnContext {
-            requirements: vec![
-                alan_agent_protocol::DelegatedCapabilityRequirement::MountRead {
-                    path: Some(PathBuf::from("/mnt/source")),
-                },
-                alan_agent_protocol::DelegatedCapabilityRequirement::Github,
-            ],
-        });
-
-        let decision = evaluate_delegated_launch_capabilities(
-            &parent,
-            &mut spec,
-            &capability_plan(Some(host_mount), &["read_file"]),
-        )
-        .await
-        .unwrap()
-        .unwrap();
-
-        assert_eq!(
-            decision.recovery,
-            alan_agent_protocol::DelegatedCapabilityRecovery::Narrowed
-        );
-        assert!(spec.launch.task.contains("NARROWED DELEGATION SCOPE"));
-        assert!(spec.launch.task.contains("Withheld capabilities: github"));
-    }
-
-    #[tokio::test]
-    async fn delegated_spawn_boundary_declines_unsatisfied_mount() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("unused"),
-        );
-        let mut spec = launch_spec(temp.path().join("agent"));
-        spec.delegated = Some(alan_agent_protocol::DelegatedSpawnContext {
-            requirements: vec![
-                alan_agent_protocol::DelegatedCapabilityRequirement::MountRead {
-                    path: Some(PathBuf::from("/mnt/private")),
-                },
-            ],
-        });
-
-        let error = evaluate_delegated_launch_capabilities(
-            &parent,
-            &mut spec,
-            &capability_plan(None, &["read_file"]),
-        )
-        .await
-        .unwrap_err();
-        let rejection = error.downcast_ref::<DelegatedSpawnRejected>().unwrap();
-
-        assert_eq!(
-            rejection.decision.recovery,
-            alan_agent_protocol::DelegatedCapabilityRecovery::AskUser
-        );
-        assert_eq!(rejection.decision.unsatisfied.len(), 1);
-    }
-
-    fn completed_response(text: &str) -> GenerationResponse {
-        GenerationResponse {
-            content: text.to_string(),
-            thinking: None,
-            thinking_signature: None,
-            redacted_thinking: Vec::new(),
-            tool_calls: Vec::new(),
-            usage: Some(TokenUsage {
-                prompt_tokens: 8,
-                cached_prompt_tokens: None,
-                completion_tokens: 4,
-                total_tokens: 12,
-                reasoning_tokens: None,
-            }),
-            finish_reason: None,
-            warnings: Vec::new(),
-            provider_response_id: None,
-            provider_response_status: None,
-        }
-    }
-
-    fn capability_view_with_package_child_agent(
-        temp: &TempDir,
-    ) -> crate::skills::ResolvedCapabilityView {
-        let package_store = temp.path().join("package-store");
-        let package_root = package_store.join("repo-review");
-        std::fs::create_dir_all(package_root.join("agents/reviewer")).unwrap();
-        std::fs::write(
-            package_root.join("SKILL.md"),
-            r#"---
-name: Repo Review
-description: Review repository changes
----
-
-Body
-"#,
-        )
-        .unwrap();
-        std::fs::write(
-            package_root.join("agents/reviewer/agent.toml"),
-            "tool_repeat_limit = 4\n",
-        )
-        .unwrap();
-        crate::skills::ResolvedCapabilityView::from_package_dirs(vec![
-            crate::skills::ScopedPackageDir {
-                path: package_store,
-                scope: crate::skills::SkillScope::Descriptor,
-            },
-        ])
-    }
-
-    fn memfs_transport() -> InProcessTransport {
-        InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::new()))
-    }
-
-    fn host_launch_root(path: impl Into<PathBuf>) -> ResolvedLaunchRoot {
-        ResolvedLaunchRoot {
-            root_dir: path.into(),
-            file_tree: None,
-        }
-    }
-
-    fn package_skill_descriptor(id: &str) -> crate::ProcessFileTree {
-        crate::ProcessFileTree::new(std::collections::BTreeMap::from([(
-            "SKILL.md".to_string(),
-            format!("---\nname: {id}\ndescription: test\n---\n").into_bytes(),
-        )]))
-        .unwrap()
-    }
-
-    fn namespace_from_child_plan(plan: &ChildNamespaceAssemblyPlan) -> KernelNamespace {
-        let mut namespace = plan.launch_context.namespace.child();
-        namespace.mount(
-            &plan.agent_mount,
-            memfs_transport(),
-            KernelAccess::ReadWrite,
-        );
-        namespace.mount(&plan.llm_mount, memfs_transport(), KernelAccess::ReadWrite);
-        namespace.mount(&plan.srv_mount, memfs_transport(), KernelAccess::ReadOnly);
-        namespace.mount(
-            &plan.route_mount,
-            memfs_transport(),
-            KernelAccess::ReadWrite,
-        );
-        for mount in &plan.bin_tool_mounts {
-            namespace.mount(mount, memfs_transport(), KernelAccess::ReadOnly);
-        }
-        for name in plan.bin_tool_names() {
-            namespace.mount(
-                &format!("/lib/exec/{name}"),
-                memfs_transport(),
-                KernelAccess::ReadOnly,
-            );
-        }
-        namespace
-    }
-
-    async fn read_proc_path(fs: &KernelProcFs, names: Vec<String>, fid: Fid) -> String {
-        fs.walk(Fid::ROOT, fid, &names).await.unwrap();
-        fs.open(fid, OpenMode::Read).await.unwrap();
-        String::from_utf8(fs.read(fid, 0, 4096).await.unwrap()).unwrap()
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_inherits_namespace_tools_but_not_optional_handles() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state_with_capability_view(
-            &temp,
-            requests.clone(),
-            response.clone(),
-            crate::skills::ResolvedCapabilityView::default(),
-        );
-        let root_dir = temp.path().join("definition");
-        let spec = launch_spec(root_dir);
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(result.output_text, "Child finished cleanly.");
-        let recorded = requests.0.lock().unwrap();
-        assert_eq!(recorded.len(), 1);
-        let request = &recorded[0];
-        assert!(request.tools.iter().any(|tool| tool.name == "alpha"));
-        assert!(request.tools.iter().any(|tool| tool.name == "beta"));
-        let user_text = request
-            .messages
-            .iter()
-            .map(|message| message.content.clone())
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(user_text.contains("Review the repository changes"));
-        assert!(!user_text.contains("Parent Conversation Snapshot"));
-        assert!(!user_text.contains("Parent Plan Snapshot"));
-        assert!(!user_text.contains("Parent Tool Results"));
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_reuses_the_passed_callable_connection() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Shared Connection completed the child.");
-        let parent = make_parent_state(&temp, requests.clone(), response);
-
-        let child = spawn_child_runtime(&parent, launch_spec(temp.path().join("definition")))
-            .await
-            .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(result.output_text, "Shared Connection completed the child.");
-        assert_eq!(requests.0.lock().unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_binds_requested_parent_handles() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Bound handles processed.");
-        let parent = make_parent_state_with_capability_view(
-            &temp,
-            requests.clone(),
-            response.clone(),
-            crate::skills::ResolvedCapabilityView::default(),
-        );
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.handles = vec![
-            SpawnHandle::ConversationSnapshot,
-            SpawnHandle::Plan,
-            SpawnHandle::ToolResults,
-        ];
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let recorded = requests.0.lock().unwrap();
-        let user_text = recorded
-            .iter()
-            .flat_map(|request| {
-                request
-                    .messages
-                    .iter()
-                    .map(|message| message.content.clone())
-            })
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(user_text.contains("Parent Conversation Snapshot"));
-        assert!(user_text.contains("Parent Plan Snapshot"));
-        assert!(user_text.contains("Parent Tool Results"));
-        assert!(user_text.contains("Inspect the changed files"));
-        assert!(user_text.contains("tool output"));
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_rejects_artifact_handle_without_runtime_binding() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Artifacts are not supported.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.handles = vec![SpawnHandle::Artifacts];
-
-        let err = match spawn_child_runtime_with_client_factory(&parent, spec, |_| unreachable!())
-            .await
-        {
-            Ok(_) => panic!("artifact handle should be rejected until artifact routing exists"),
-            Err(err) => err,
-        };
-
-        assert!(
-            err.to_string()
-                .contains("Child Agent Process launches do not support artifact routing yet")
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_rejects_output_dir_without_runtime_binding() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Artifacts are not supported.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.launch.output_dir = Some(temp.path().join("repo/out"));
-
-        let err = match spawn_child_runtime_with_client_factory(&parent, spec, |_| unreachable!())
-            .await
-        {
-            Ok(_) => panic!("output_dir should be rejected until artifact routing exists"),
-            Err(err) => err,
-        };
-
-        assert!(
-            err.to_string()
-                .contains("Child Agent Process launches do not support artifact routing yet")
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_filters_namespace_tools_with_override() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Only one tool should be visible.");
-        let parent = make_parent_state_with_capability_view(
-            &temp,
-            requests.clone(),
-            response.clone(),
-            crate::skills::ResolvedCapabilityView::default(),
-        );
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let recorded = requests.0.lock().unwrap();
-        let tool_names = recorded[0]
-            .tools
-            .iter()
-            .map(|tool| tool.name.as_str())
-            .collect::<Vec<_>>();
-        assert!(tool_names.contains(&"alpha"));
-        assert!(!tool_names.contains(&"beta"));
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_respects_empty_namespace_tool_override() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("No tools should be visible.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: Vec::new(),
-        });
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let recorded = requests.0.lock().unwrap();
-        let tool_names = recorded[0]
-            .tools
-            .iter()
-            .map(|tool| tool.name.as_str())
-            .collect::<Vec<_>>();
-        assert!(!tool_names.contains(&"alpha"));
-        assert!(!tool_names.contains(&"beta"));
-    }
-
-    #[tokio::test]
-    async fn child_namespace_plan_mounts_only_allowed_tools() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-
-        let launch_context = parent
-            .namespace_environment()
-            .launch_context()
-            .unwrap()
-            .child();
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            launch_context,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(plan.llm_mount, "/mnt/llm");
-        assert_eq!(plan.llm_connection_name().unwrap(), "default");
-        assert_eq!(plan.srv_mount, "/srv");
-        assert_eq!(plan.route_mount, "/mnt/route");
-        assert_eq!(plan.cwd, Some(PathBuf::from("/mnt/source")));
-        assert_eq!(plan.launch_context.cwd, "/mnt/source");
-        assert_eq!(plan.bin_tool_mounts, vec!["/bin/alpha"]);
-    }
-
-    #[tokio::test]
-    async fn child_clone_exec_spec_declares_agent_and_llm_mounts_for_pid() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut child_core_config = parent.core_config.clone();
-        child_core_config.connection_profile = Some("child-main".to_string());
-        let spec = launch_spec(root_dir);
-
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &child_core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let exec = plan.clone_exec_spec_for_pid("42", "/bin/alan-agent", ["--boot"]);
-
-        assert_eq!(
-            serde_json::to_value(&exec).unwrap(),
-            json!({
-                "executable": "/bin/alan-agent",
-                "args": ["--boot"],
-                "descriptors": {
-                    "3": "/agent-definition",
-                    "4": "/memory"
-                },
-                "namespace": {
-                    "mounts": [
-                        {"path": "/agent", "access": "rw"},
-                        {"path": "/agent-definition", "access": "ro"},
-                        {"path": "/bin/alpha", "access": "ro"},
-                        {"path": "/bin/beta", "access": "ro"},
-                        {"path": "/lib/exec/alpha", "access": "ro"},
-                        {"path": "/lib/exec/beta", "access": "ro"},
-                        {"path": "/memory", "access": "rw"},
-                        {"path": "/mnt/llm", "access": "rw"},
-                        {"path": "/mnt/route", "access": "rw"},
-                        {"path": "/mnt/source", "access": "rw"},
-                        {"path": "/srv", "access": "ro"}
-                    ]
-                }
-            })
-        );
-        let decoded: ExecSpec = serde_json::from_value(serde_json::to_value(&exec).unwrap())
-            .expect("child clone document uses the kernel ExecSpec contract");
-        assert_eq!(decoded, exec);
-    }
-
-    #[tokio::test]
-    async fn child_clone_exec_spec_declares_only_allowed_bin_mounts() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let manifest = plan.namespace_manifest_for_pid("99");
-
-        assert_eq!(
-            serde_json::to_value(&manifest).unwrap(),
-            json!({
-                "mounts": [
-                    {"path": "/agent", "access": "rw"},
-                    {"path": "/agent-definition", "access": "ro"},
-                    {"path": "/bin/alpha", "access": "ro"},
-                    {"path": "/lib/exec/alpha", "access": "ro"},
-                    {"path": "/memory", "access": "rw"},
-                    {"path": "/mnt/llm", "access": "rw"},
-                    {"path": "/mnt/route", "access": "rw"},
-                    {"path": "/mnt/source", "access": "rw"},
-                    {"path": "/srv", "access": "ro"}
-                ]
-            })
-        );
-    }
-
-    #[tokio::test]
-    async fn child_clone_exec_spec_commits_through_proc_clone_with_planned_namespace() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let procfs = KernelProcFs::new();
-        let spawner = procfs.for_spawner(
-            None,
-            namespace_from_child_plan(&plan),
-            KernelCredentials::user("alan"),
-        );
-
-        spawner
-            .walk(Fid::ROOT, Fid(90), &["clone".to_string()])
-            .await
-            .unwrap();
-        spawner.open(Fid(90), OpenMode::ReadWrite).await.unwrap();
-        let pid = String::from_utf8(spawner.read(Fid(90), 0, 64).await.unwrap()).unwrap();
-        let exec = plan.clone_exec_spec_for_pid(&pid, "/bin/alan-agent", ["--boot"]);
-        let exec_bytes = serde_json::to_vec(&exec).unwrap();
-        spawner.write(Fid(90), 0, &exec_bytes).await.unwrap();
-        spawner.clunk(Fid(90)).await.unwrap();
-
-        let namespace =
-            read_proc_path(&procfs, vec![pid.clone(), "namespace".to_string()], Fid(91)).await;
-        assert!(
-            namespace.lines().any(|line| line == "/agent rw"),
-            "agent overlay is mounted at /agent: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/mnt/llm rw"),
-            "llm connection is mounted: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/bin/alpha ro"),
-            "allowed tool executable is mounted read-only: {namespace:?}"
-        );
-        assert!(
-            !namespace.lines().any(|line| line.contains("<child-pid>")),
-            "placeholder is expanded before the process becomes public: {namespace:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn child_namespace_launch_and_supervisor_reattachment_use_proc_pid_files() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let launch_procfs = KernelProcFs::new();
-        let tool_runner =
-            crate::tools::ToolProcessRunner::from_registry(&parent_test_tools(&parent.core_config));
-        let runtime_procfs = launch_procfs
-            .clone()
-            .with_runner(Arc::new(tool_runner.clone()));
-        let handles = ChildNamespaceLaunchHandles::new(
-            Arc::new(alan_agentfs::AgentFs::new()),
-            memfs_transport(),
-            memfs_transport(),
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/alpha",
-            memfs_transport(),
-            "/lib/exec/alpha",
-            memfs_transport(),
-        );
-
-        let launch = spawn_child_namespace_runtime_environment(
-            &launch_procfs,
-            &runtime_procfs,
-            &plan,
-            handles,
-            None,
-            tool_runner.clone(),
-            plan.execution_binding(temp.path().join("scratch")).unwrap(),
-            None,
-            "/bin/alan-agent",
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(launch.pid, "1");
-        assert_eq!(launch.environment.agent_path(), "/agent/1");
-        assert_eq!(launch.environment.llm_connection(), "default");
-        assert_eq!(
-            launch.exec,
-            plan.clone_exec_spec_for_pid("1", "/bin/alan-agent", std::iter::empty::<String>())
-        );
-
-        assert_eq!(
-            read_proc_path(
-                &launch_procfs,
-                vec![launch.pid.clone(), "status".to_string()],
-                Fid(90),
-            )
-            .await,
-            "running\n",
-            "child agent process should stay running after launch"
-        );
-        let namespace = read_proc_path(
-            &launch_procfs,
-            vec![launch.pid.clone(), "namespace".to_string()],
-            Fid(92),
-        )
-        .await;
-        assert!(
-            namespace.lines().any(|line| line == "/agent rw"),
-            "agent overlay is mounted: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/mnt/llm rw"),
-            "llm connection is present: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/mnt/route rw"),
-            "routefs tree is present: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/srv ro"),
-            "service handle registry is present: {namespace:?}"
-        );
-        assert!(
-            namespace.lines().any(|line| line == "/bin/alpha ro"),
-            "allowed tool mount is present: {namespace:?}"
-        );
-
-        let child_handles = ChildNamespaceLaunchHandles::new(
-            Arc::new(alan_agentfs::AgentFs::new()),
-            memfs_transport(),
-            memfs_transport(),
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/alpha",
-            memfs_transport(),
-            "/lib/exec/alpha",
-            memfs_transport(),
-        );
-        let nested = spawn_child_namespace_runtime_environment(
-            &launch_procfs,
-            &runtime_procfs,
-            &plan,
-            child_handles,
-            Some(TestParentProcessContext {
-                agent_root: launch.agent_root.clone(),
-                pid: alan_kernel::Pid(launch.pid.parse().unwrap()),
-            }),
-            tool_runner.clone(),
-            plan.execution_binding(temp.path().join("scratch")).unwrap(),
-            None,
-            "/bin/alan-agent",
-        )
-        .await
-        .unwrap();
-        assert_eq!(nested.pid, "2");
-        assert_eq!(
-            read_proc_path(
-                &launch_procfs,
-                vec![nested.pid.clone(), "parent".to_string()],
-                Fid(94),
-            )
-            .await,
-            "1"
-        );
-        let parent_shell = alan_shell::Shell::new(launch.environment.root_transport());
-        assert_eq!(
-            parent_shell.ls("/agent/1/children").await.unwrap(),
-            vec![nested.pid.clone()],
-            "delegated Agent Process must be inspectable from the parent AgentFS view"
-        );
-        record_child_launch_failure_process(
-            &nested.lifecycle,
-            &anyhow::anyhow!("simulated child runtime startup failure"),
-        )
-        .await;
-        assert_eq!(
-            nested
-                .environment
-                .read_process_exit_code(&nested.pid)
-                .await
-                .unwrap(),
-            Some(1)
-        );
-        assert!(
-            parent_shell
-                .ls("/agent/1/children")
-                .await
-                .unwrap()
-                .is_empty(),
-            "failed child launch must leave no running child entry"
-        );
-
-        let tool = launch
-            .environment
-            .run_tool_action("alpha", "/bin/alpha", ["{}"])
-            .await
-            .unwrap();
-        assert_eq!(tool.pid, "3");
-        assert_eq!(tool.action_id, "a0");
-        assert_eq!(tool.output.trim(), r#"{"ok":true}"#);
-        let tool_namespace = read_proc_path(
-            &launch_procfs,
-            vec![tool.pid.clone(), "namespace".to_string()],
-            Fid(93),
-        )
-        .await;
-        assert!(
-            tool_namespace.lines().any(|line| line == "/agent rw"),
-            "child-spawned processes inherit the agent overlay: {tool_namespace:?}"
-        );
-        assert!(
-            tool_namespace.lines().any(|line| line == "/mnt/llm rw"),
-            "child-spawned processes inherit the llm connection: {tool_namespace:?}"
-        );
-        assert!(
-            tool_namespace.lines().any(|line| line == "/mnt/route rw"),
-            "child-spawned processes inherit routefs: {tool_namespace:?}"
-        );
-        assert!(
-            tool_namespace.lines().any(|line| line == "/srv ro"),
-            "child-spawned processes inherit /srv handles: {tool_namespace:?}"
-        );
-        assert!(
-            tool_namespace.lines().any(|line| line == "/bin/alpha ro"),
-            "child-spawned processes inherit mounted tools: {tool_namespace:?}"
-        );
-
-        let process_reader = launch.environment.clone();
-        let process_pid = launch.pid.clone();
-        launch
-            .environment
-            .write_assistant_output("AgentFS child result")
-            .await
-            .unwrap();
-        crate::runtime::ui_surfaces::turn_completed(&launch.environment, false)
-            .await
-            .unwrap();
-        let controller = DelegatedChildRunSupervisor::new(DelegatedChildRunSupervision {
-            runtime: None,
-            startup_metadata: test_startup_metadata("child-machine", None, false),
-            child_run_id: format!("test-child-run-{}", uuid::Uuid::new_v4()),
-            child_run_registry: ChildRunRegistry::default(),
-            timeout: None,
-            process_lifecycle: launch.lifecycle,
-            process_environment: launch.environment,
-            process_pid: process_pid.clone(),
-        });
-
-        let result = controller.join().await.unwrap();
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(result.output_text, "AgentFS child result");
-        assert_eq!(
-            process_reader
-                .read_process_exit_code(&process_pid)
-                .await
-                .unwrap(),
-            Some(0),
-            "normal completion must not be rewritten as ctl cancellation (130)"
-        );
-    }
-
-    #[tokio::test]
-    async fn external_proc_ctl_stops_child_runtime_controller() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child should be stopped externally.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.tool_profile = Some(alan_agent_protocol::SpawnToolProfileOverride {
-            allowed_tools: vec!["alpha".to_string()],
-        });
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let launch_procfs = KernelProcFs::new();
-        let tool_runner =
-            crate::tools::ToolProcessRunner::from_registry(&parent_test_tools(&parent.core_config));
-        let runtime_procfs = launch_procfs
-            .clone()
-            .with_runner(Arc::new(tool_runner.clone()));
-        let handles = ChildNamespaceLaunchHandles::new(
-            Arc::new(alan_agentfs::AgentFs::new()),
-            memfs_transport(),
-            memfs_transport(),
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/alpha",
-            memfs_transport(),
-            "/lib/exec/alpha",
-            memfs_transport(),
-        );
-        let launch = spawn_child_namespace_runtime_environment(
-            &launch_procfs,
-            &runtime_procfs,
-            &plan,
-            handles,
-            None,
-            tool_runner,
-            plan.execution_binding(temp.path().join("scratch")).unwrap(),
-            None,
-            "/bin/alan-agent",
-        )
-        .await
-        .unwrap();
-        let process_pid = launch.pid.clone();
-        let process_environment = launch.environment.clone();
-        let controller = DelegatedChildRunSupervisor::new(DelegatedChildRunSupervision {
-            runtime: None,
-            startup_metadata: test_startup_metadata("child-machine", None, false),
-            child_run_id: format!("test-child-run-{}", uuid::Uuid::new_v4()),
-            child_run_registry: ChildRunRegistry::default(),
-            timeout: None,
-            process_lifecycle: launch.lifecycle,
-            process_environment: launch.environment,
-            process_pid: process_pid.clone(),
-        });
-
-        assert_eq!(process_environment.ui_events_offset().await.unwrap(), 0);
-        process_environment
-            .write_process_control_for_pid(&process_pid, "cancel")
-            .await
-            .unwrap();
-        let result = tokio::time::timeout(Duration::from_secs(2), controller.join())
-            .await
-            .expect("controller must observe external proc cancellation")
-            .unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Terminated);
-        assert!(
-            result
-                .error_message
-                .as_deref()
-                .is_some_and(|message| message.contains("/proc/<pid>/ctl"))
-        );
-        assert_eq!(
-            process_environment
-                .read_process_exit_code(&process_pid)
-                .await
-                .unwrap(),
-            Some(130)
-        );
-    }
-
-    #[tokio::test]
-    async fn child_namespace_launch_attaches_mount_grant_applicator_factory() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let spec = launch_spec(root_dir);
-        let mut plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        plan.launch_context
-            .host_mounts
-            .retain(|grant| grant.namespace_path != "/agent-definition");
-        let package_child_definition = temp.path().join("package-child-definition");
-        plan.launch_context.host_mounts.push(
-            crate::HostMountGrant::new(
-                "/agent-definition",
-                package_child_definition.clone(),
-                KernelAccess::ReadOnly,
-            )
-            .unwrap(),
-        );
-        let launch_procfs = KernelProcFs::new();
-        let tool_runner =
-            crate::tools::ToolProcessRunner::from_registry(&parent_test_tools(&parent.core_config));
-        let runtime_procfs = launch_procfs
-            .clone()
-            .with_runner(Arc::new(tool_runner.clone()));
-        let handles = ChildNamespaceLaunchHandles::new(
-            Arc::new(alan_agentfs::AgentFs::new()),
-            memfs_transport(),
-            memfs_transport(),
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/alpha",
-            memfs_transport(),
-            "/lib/exec/alpha",
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/beta",
-            memfs_transport(),
-            "/lib/exec/beta",
-            memfs_transport(),
-        );
-        let factory = Arc::new(RecordingMountGrantApplicatorFactory::default());
-
-        let mut launch = spawn_child_namespace_runtime_environment(
-            &launch_procfs,
-            &runtime_procfs,
-            &plan,
-            handles,
-            None,
-            tool_runner,
-            plan.execution_binding(temp.path().join("scratch")).unwrap(),
-            Some(factory.clone()),
-            "/bin/alan-agent",
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(factory.created_count(), 1);
-        assert_eq!(
-            factory.applied_grants(),
-            [ApprovedMountGrant::new(
-                "/agent-definition",
-                package_child_definition,
-                ApprovedMountGrantAccess::ReadOnly,
-                "Agent Definition launch reference",
-            )]
-        );
-        let definition_namespace = read_proc_path(
-            &launch_procfs,
-            vec![launch.pid.clone(), "namespace".to_string()],
-            Fid(95),
-        )
-        .await;
-        assert!(
-            definition_namespace
-                .lines()
-                .any(|line| line == "/agent-definition ro"),
-            "child Agent Process must receive its target definition: {definition_namespace:?}"
-        );
-        let applied = launch
-            .environment
-            .apply_approved_mount_grant(&ApprovedMountGrant::new(
-                "/mnt/project",
-                PathBuf::from("/unused/by/test/applicator"),
-                ApprovedMountGrantAccess::ReadWrite,
-                "Need project files",
-            ));
-        assert!(applied.namespace_applied);
-        assert_eq!(applied.namespace_error, None);
-
-        let namespace = read_proc_path(
-            &launch_procfs,
-            vec![launch.pid.clone(), "namespace".to_string()],
-            Fid(94),
-        )
-        .await;
-        assert!(
-            namespace.lines().any(|line| line == "/mnt/project rw"),
-            "child Agent Process namespace should reflect applicator live mounts: {namespace:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn child_namespace_launch_handles_share_parent_routefs() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let routefs = Arc::new(alan_routefs::RouteFs::new());
-        routefs
-            .install_rule(
-                "10-results",
-                alan_routefs::RuleSpec::for_type("result", "review"),
-            )
-            .await
-            .unwrap();
-        let mut parent = make_parent_state(&temp, requests, response);
-        let launch_context = inherited_launch_context(&parent);
-        parent.environment = namespace_environment_for_parent_test_with_route(routefs.clone())
-            .with_launch_context(launch_context);
-
-        let root_dir = temp.path().join("definition");
-        let spec = launch_spec(root_dir);
-        let plan = build_child_namespace_assembly_plan(
-            &parent,
-            &spec,
-            &parent.core_config,
-            inherited_launch_context(&parent),
-        )
-        .await
-        .unwrap();
-        let launch_procfs = KernelProcFs::new();
-        let runtime_procfs = launch_procfs.clone().with_runner(Arc::new(
-            crate::tools::ToolProcessRunner::from_registry(&parent_test_tools(&parent.core_config)),
-        ));
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        llmfs.register_connection(
-            &plan.llm_connection_name().unwrap(),
-            Box::new(ChildLlmProvider::new(LlmClient::new(
-                RecordingProvider::new(RecordedRequests::default(), completed_response("unused")),
-            ))),
-        );
-        let handles = ChildNamespaceLaunchHandles::new(
-            Arc::new(alan_agentfs::AgentFs::new()),
-            InProcessTransport::new(llmfs),
-            memfs_transport(),
-            InProcessTransport::new(routefs.clone()),
-        )
-        .with_tool_package(
-            "/bin/alpha",
-            memfs_transport(),
-            "/lib/exec/alpha",
-            memfs_transport(),
-        )
-        .with_tool_package(
-            "/bin/beta",
-            memfs_transport(),
-            "/lib/exec/beta",
-            memfs_transport(),
-        );
-
-        let launch = spawn_child_namespace_runtime_environment(
-            &launch_procfs,
-            &runtime_procfs,
-            &plan,
-            handles,
-            None,
-            crate::tools::ToolProcessRunner::from_registry(&parent_test_tools(&parent.core_config)),
-            plan.execution_binding(temp.path().join("scratch")).unwrap(),
-            None,
-            "/bin/alan-agent",
-        )
-        .await
-        .unwrap();
-
-        let child_shell = alan_shell::Shell::new(launch.environment.root_transport());
-        let message = serde_json::to_vec(&json!({
-            "version": 1,
-            "type": "result",
-            "content": "child result"
-        }))
-        .unwrap();
-        child_shell
-            .write("/mnt/route/send", &message)
-            .await
-            .unwrap();
-
-        let parent_route_shell = alan_shell::Shell::new(InProcessTransport::new(routefs));
-        let routed =
-            String::from_utf8(parent_route_shell.cat("/ports/review").await.unwrap()).unwrap();
-        assert!(routed.contains(r#""type":"result""#), "{routed}");
-        assert!(routed.contains(r#""content":"child result""#), "{routed}");
-    }
-
-    #[tokio::test]
-    async fn child_tool_runner_rejects_unmounted_tool_executables() {
-        let mut child_tools = ToolRegistry::new();
-        child_tools.register(MarkerTool::new("alpha", "mounted-only"));
-        let runner = crate::tools::ToolProcessRunner::from_registry(&child_tools);
-        let invocation = alan_kernel::ProcessInvocation {
-            pid: alan_kernel::Pid(1),
-            parent: Some(alan_kernel::Pid(0)),
-            credentials: alan_kernel::Credentials::user("child-agent"),
-            namespace: alan_kernel::Namespace::new(),
-            exec: alan_kernel::ExecSpec {
-                executable: "/bin/alpha".to_string(),
-                args: vec!["{}".to_string()],
-                namespace: None,
-                descriptors: Default::default(),
-            },
-        };
-
-        let outcome = alan_kernel::ProcessRunner::run(&runner, invocation).await;
-
-        assert_eq!(outcome.exit_code, 127);
-        assert_eq!(outcome.output, b"executable is not mounted\n");
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_conversation_snapshot_excludes_tool_outputs_without_handle() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Snapshot captured.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.handles = vec![SpawnHandle::ConversationSnapshot];
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let recorded = requests.0.lock().unwrap();
-        let user_text = recorded
-            .iter()
-            .flat_map(|request| {
-                request
-                    .messages
-                    .iter()
-                    .map(|message| message.content.clone())
-            })
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(user_text.contains("Parent Conversation Snapshot"));
-        assert!(!user_text.contains("tool output"));
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_uses_effective_launch_root_config_for_llm_setup() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let root_dir = temp.path().join("definition");
-        std::fs::write(
-            root_dir.join("agent.toml"),
-            r#"
-tool_repeat_limit = 9
-"#,
-        )
-        .unwrap();
-        let seen_config = Arc::new(Mutex::new(None::<crate::Config>));
-        let seen_config_for_factory = seen_config.clone();
-
-        let child =
-            spawn_child_runtime_with_client_factory(&parent, launch_spec(root_dir), |config| {
-                *seen_config_for_factory.lock().unwrap() = Some(config.clone());
-                Ok(LlmClient::new(RecordingProvider::new(
-                    requests.clone(),
-                    response.clone(),
-                )))
-            })
-            .await
-            .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let seen_config = seen_config.lock().unwrap().clone().unwrap();
-        assert_eq!(seen_config.effective_model(), "gpt-5.4");
-        assert_eq!(seen_config.tool_repeat_limit, 9);
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_preserves_explicit_connection_profile() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child used the explicit profile.");
-        let mut parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let profile_id = "explicit-main";
-        parent.core_config.connection_profile = Some(profile_id.to_string());
-        let launch_context = parent
-            .namespace_environment()
-            .launch_context()
-            .unwrap()
-            .clone();
-        parent.environment = namespace_environment_for_parent_test_with_connection(
-            Arc::new(alan_routefs::RouteFs::new()),
-            Arc::new(alan_llmfs::LlmFs::new()),
-            profile_id,
-        )
-        .with_launch_context(launch_context);
-        let root_dir = temp.path().join("definition");
-        let seen_config = Arc::new(Mutex::new(None::<crate::Config>));
-        let seen_config_for_factory = seen_config.clone();
-
-        let child =
-            spawn_child_runtime_with_client_factory(&parent, launch_spec(root_dir), |config| {
-                *seen_config_for_factory.lock().unwrap() = Some(config.clone());
-                Ok(LlmClient::new(RecordingProvider::new(
-                    requests.clone(),
-                    response.clone(),
-                )))
-            })
-            .await
-            .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(
-            seen_config
-                .lock()
-                .unwrap()
-                .as_ref()
-                .and_then(|config| config.connection_profile.as_deref()),
-            Some(profile_id)
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_rejects_unpassed_definition_connection_reference() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child used its definition profile.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let profile_id = "child-main";
-        let root_dir = temp.path().join("definition");
-        std::fs::write(
-            root_dir.join("agent.toml"),
-            format!("connection_profile = \"{profile_id}\"\n"),
-        )
-        .unwrap();
-        let error =
-            match spawn_child_runtime_with_client_factory(&parent, launch_spec(root_dir), |_| {
-                unreachable!("an unpassed Connection must fail before provider setup")
-            })
-            .await
-            {
-                Ok(_) => panic!("unpassed child Connection should be rejected"),
-                Err(error) => error,
-            };
-
-        assert!(
-            error
-                .to_string()
-                .contains("was not passed to the child Agent Process by the parent Process")
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_applies_reasoning_effort_override_after_overlay() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let root_dir = temp.path().join("definition");
-        std::fs::write(
-            root_dir.join("agent.toml"),
-            r#"
-model_reasoning_effort = "high"
-"#,
-        )
-        .unwrap();
-        let seen_config = Arc::new(Mutex::new(None::<crate::Config>));
-        let seen_config_for_factory = seen_config.clone();
-        let mut spec = launch_spec(root_dir);
-        spec.runtime_overrides.model_reasoning_effort =
-            Some(alan_agent_protocol::ReasoningEffort::Low);
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |config| {
-            *seen_config_for_factory.lock().unwrap() = Some(config.clone());
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        let seen_config = seen_config.lock().unwrap().clone().unwrap();
-        assert_eq!(
-            crate::resolve_runtime_request_controls(
-                &seen_config,
-                crate::provider_capabilities_for_config(&seen_config),
-                crate::RequestControlIntent::default(),
-            )
-            .unwrap()
-            .reasoning_effort(),
-            Some(alan_agent_protocol::ReasoningEffort::Low)
-        );
-
-        let recorded = requests.0.lock().unwrap();
-        assert_eq!(
-            recorded[0].reasoning.effort,
-            Some(alan_agent_protocol::ReasoningEffort::Low)
-        );
-    }
-
-    #[test]
-    fn child_agent_config_requires_memory_handle_for_memory_dir() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Child finished cleanly.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-
-        let mut approval_spec = launch_spec(root_dir.clone());
-        approval_spec.handles = vec![SpawnHandle::ApprovalScope];
-        let approval_config = build_child_agent_config(&parent, &approval_spec);
-        assert_eq!(approval_config.core_config.memory.store_dir, None);
-
-        let mut override_spec = launch_spec(root_dir);
-        override_spec.runtime_overrides.policy_path = Some("policy.yaml".to_string());
-        let override_config = build_child_agent_config(&parent, &override_spec);
-        assert_eq!(override_config.core_config.memory.store_dir, None);
-    }
-
-    #[test]
-    fn child_launch_contract_rejects_relative_namespace_cwd() {
-        let mut spec = launch_spec(PathBuf::from("/tmp/definition"));
-        spec.launch.cwd = Some(PathBuf::from("docs"));
-
-        let err = validate_child_launch_contract(&spec).unwrap_err();
-        assert!(
-            format!("{err:#}").contains("absolute"),
-            "expected absolute-path validation error, got {err:#}"
-        );
-    }
-
-    #[test]
-    fn child_launch_contract_rejects_non_normal_namespace_cwd() {
-        for cwd in ["/mnt/source/../other", "/mnt/./source"] {
-            let mut spec = launch_spec(PathBuf::from("/tmp/definition"));
-            spec.launch.cwd = Some(PathBuf::from(cwd));
-
-            let err = validate_child_launch_contract(&spec).unwrap_err();
-            assert!(
-                err.to_string()
-                    .contains("Invalid child Agent Process launch cwd"),
-                "expected normal-path validation error for {cwd}, got {err:#}"
-            );
-        }
-    }
-
-    #[test]
-    fn child_launch_context_does_not_inherit_parent_host_mounts_or_descriptors_by_default() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("done"),
-        );
-        let parent_context = parent.namespace_environment().launch_context().unwrap();
-        let definition = temp.path().join("child-definition");
-        let mut spec = launch_spec(definition.clone());
-        spec.handles.clear();
-
-        let definition = host_launch_root(definition);
-        let child =
-            build_child_launch_context(parent_context, &spec, None, Some(&definition)).unwrap();
-
-        assert_eq!(child.cwd, "/");
-        assert!(child.namespace.resolve("/mnt/source").is_err());
-        assert!(
-            child
-                .host_mounts
-                .iter()
-                .all(|grant| grant.namespace_path != "/mnt/source")
-        );
-        assert_eq!(
-            child
-                .descriptors
-                .keys()
-                .map(String::as_str)
-                .collect::<Vec<_>>(),
-            vec![crate::AGENT_DEFINITION_DESCRIPTOR]
-        );
-    }
-
-    #[test]
-    fn child_launch_context_keeps_package_projection_with_inherited_reference() {
-        let package = TempDir::new().unwrap();
-        let skill_root = package.path().join("skills/inherited");
-        std::fs::create_dir_all(&skill_root).unwrap();
-        std::fs::write(
-            skill_root.join("SKILL.md"),
-            "---\nname: inherited\ndescription: Inherited package Skill.\n---\n",
-        )
-        .unwrap();
-
-        let mut namespace = KernelNamespace::new();
-        namespace.mount(
-            "/lib/pkg/parent-pack",
-            memfs_transport(),
-            KernelAccess::ReadOnly,
-        );
-        let package_reference = crate::ProcessPackageReference::new(
-            "parent-pack",
-            "a".repeat(64),
-            crate::ProcessPackageKind::Installed,
-            "/lib/pkg/parent-pack",
-            vec![
-                crate::ProcessPackageSkillReference::new(
-                    "inherited",
-                    "skills/inherited",
-                    Vec::new(),
-                    package_skill_descriptor("inherited"),
-                )
-                .unwrap(),
-            ],
-            memfs_transport(),
-        )
-        .unwrap();
-        let parent = crate::ProcessLaunchContext::new(
-            namespace,
-            KernelCredentials::user("parent-agent"),
-            "/lib/pkg/parent-pack",
-        )
-        .unwrap()
-        .with_package_reference(package_reference);
-        let mut spec = launch_spec(package.path().join("unused-definition"));
-        spec.handles.clear();
-
-        let child = build_child_launch_context(&parent, &spec, None, None).unwrap();
-
-        assert_eq!(child.cwd, "/lib/pkg/parent-pack");
-        assert_eq!(child.package_references.len(), 1);
-        assert_eq!(child.package_references[0].package_id, "parent-pack");
-        assert!(child.namespace.resolve("/lib/pkg/parent-pack").is_ok());
-        assert!(child.host_mounts.is_empty());
-        let resolved = crate::ResolvedAgentDefinition::from_launch_context(
-            &child,
-            &[],
-            crate::ConfigSourceKind::Default,
-        )
-        .unwrap();
-        assert_eq!(resolved.capability_view.packages.len(), 1);
-        assert_eq!(
-            resolved.capability_view.packages[0].id,
-            "installed:parent-pack:inherited"
-        );
-    }
-
-    #[tokio::test]
-    async fn package_child_definition_is_passed_by_descriptor_and_package_mount() {
-        let parent_definition = TempDir::new().unwrap();
-        let mut namespace = KernelNamespace::new();
-        namespace.mount("/lib/pkg/review", memfs_transport(), KernelAccess::ReadOnly);
-        namespace.mount(
-            "/lib/agents/root",
-            memfs_transport(),
-            KernelAccess::ReadOnly,
-        );
-        let parent = crate::ProcessLaunchContext::new(
-            namespace,
-            KernelCredentials::user("parent-agent"),
-            "/",
-        )
-        .unwrap()
-        .with_host_mount(
-            crate::HostMountGrant::new(
-                "/lib/agents/root",
-                parent_definition.path(),
-                KernelAccess::ReadOnly,
-            )
-            .unwrap(),
-        )
-        .with_descriptor(
-            crate::AGENT_DEFINITION_DESCRIPTOR,
-            crate::ProcessDescriptor::new("/lib/agents/root").unwrap(),
-        )
-        .with_package_reference(
-            crate::ProcessPackageReference::new(
-                "review",
-                "a".repeat(64),
-                crate::ProcessPackageKind::Installed,
-                "/lib/pkg/review",
-                Vec::new(),
-                memfs_transport(),
-            )
-            .unwrap(),
-        );
-        let definition = ResolvedLaunchRoot {
-            root_dir: PathBuf::from("/lib/pkg/review/skills/review/agents/critic"),
-            file_tree: Some(
-                crate::ProcessFileTree::new(std::collections::BTreeMap::from([(
-                    "agent.toml".to_string(),
-                    b"tool_repeat_limit = 3\n".to_vec(),
-                )]))
-                .unwrap(),
-            ),
-        };
-        let mut spec = launch_spec(definition.root_dir.clone());
-        spec.handles.clear();
-
-        let child = build_child_launch_context(&parent, &spec, None, Some(&definition)).unwrap();
-
-        assert!(child.host_mounts.is_empty());
-        assert!(child.namespace.union_at("/lib/agents/root").is_empty());
-        let descriptor = child
-            .descriptor(crate::AGENT_DEFINITION_DESCRIPTOR)
-            .unwrap();
-        assert_eq!(descriptor.path, definition.root_dir.to_string_lossy());
-        assert!(
-            descriptor
-                .file_tree
-                .as_ref()
-                .is_some_and(|tree| tree.contains_file("agent.toml"))
-        );
-
-        let mut plan = capability_plan(None, &[]);
-        plan.launch_context = child;
-        let manifest = plan.namespace_manifest_for_pid("99");
-        assert!(manifest.mounts.iter().any(|mount| {
-            mount.path == "/lib/pkg/review" && mount.access == ExecNamespaceAccess::ReadOnly
-        }));
-        assert!(
-            !manifest
-                .mounts
-                .iter()
-                .any(|mount| { mount.path == "/lib/pkg/review/skills/review/agents/critic" })
-        );
-
-        let procfs = KernelProcFs::new();
-        let spawner = procfs.for_spawner(
-            None,
-            namespace_from_child_plan(&plan),
-            KernelCredentials::user("alan"),
-        );
-        spawner
-            .walk(Fid::ROOT, Fid(90), &["clone".to_string()])
-            .await
-            .unwrap();
-        spawner.open(Fid(90), OpenMode::ReadWrite).await.unwrap();
-        let pid = String::from_utf8(spawner.read(Fid(90), 0, 64).await.unwrap()).unwrap();
-        let exec = plan.clone_exec_spec_for_pid(&pid, "/bin/alan-agent", ["--boot"]);
-        spawner
-            .write(Fid(90), 0, &serde_json::to_vec(&exec).unwrap())
-            .await
-            .unwrap();
-        spawner.clunk(Fid(90)).await.unwrap();
-    }
-
-    #[test]
-    fn child_launch_context_binds_a_noncanonical_parent_definition_path() {
-        let definition = TempDir::new().unwrap();
-        let mut namespace = KernelNamespace::new();
-        namespace.mount(
-            "/lib/agents/root",
-            memfs_transport(),
-            KernelAccess::ReadOnly,
-        );
-        let parent = crate::ProcessLaunchContext::new(
-            namespace,
-            KernelCredentials::user("parent-agent"),
-            "/",
-        )
-        .unwrap()
-        .with_host_mount(
-            crate::HostMountGrant::new(
-                "/lib/agents/root",
-                definition.path(),
-                KernelAccess::ReadOnly,
-            )
-            .unwrap(),
-        )
-        .with_descriptor(
-            crate::AGENT_DEFINITION_DESCRIPTOR,
-            crate::ProcessDescriptor::new("/lib/agents/root").unwrap(),
-        );
-        let spec = launch_spec(definition.path().to_path_buf());
-
-        let definition_root = host_launch_root(definition.path());
-        let child =
-            build_child_launch_context(&parent, &spec, None, Some(&definition_root)).unwrap();
-
-        assert!(child.namespace.resolve("/agent-definition").is_ok());
-        assert_eq!(
-            child
-                .descriptor(crate::AGENT_DEFINITION_DESCRIPTOR)
-                .unwrap()
-                .path,
-            "/agent-definition"
-        );
-    }
-
-    #[test]
-    fn child_launch_context_keeps_bootstrap_definition_until_descendant_target_is_projected() {
-        let root = TempDir::new().unwrap();
-        let parent_definition = root.path().join("parent-definition");
-        let package_root = root.path().join("package");
-        let child_definition = package_root.join("agents/reviewer");
-        std::fs::create_dir_all(&parent_definition).unwrap();
-        std::fs::create_dir_all(&child_definition).unwrap();
-
-        let mut namespace = KernelNamespace::new();
-        namespace.mount(
-            "/agent-definition",
-            memfs_transport(),
-            KernelAccess::ReadOnly,
-        );
-        namespace.mount("/mnt/package", memfs_transport(), KernelAccess::ReadOnly);
-        let parent = crate::ProcessLaunchContext::new(
-            namespace,
-            KernelCredentials::user("parent-agent"),
-            "/",
-        )
-        .unwrap()
-        .with_host_mount(
-            crate::HostMountGrant::new(
-                "/agent-definition",
-                parent_definition,
-                KernelAccess::ReadOnly,
-            )
-            .unwrap(),
-        )
-        .with_host_mount(
-            crate::HostMountGrant::new("/mnt/package", package_root, KernelAccess::ReadOnly)
-                .unwrap(),
-        )
-        .with_descriptor(
-            crate::AGENT_DEFINITION_DESCRIPTOR,
-            crate::ProcessDescriptor::new("/agent-definition").unwrap(),
-        );
-
-        let child_definition_root = host_launch_root(&child_definition);
-        let child = build_child_launch_context(
-            &parent,
-            &launch_spec(child_definition.clone()),
-            None,
-            Some(&child_definition_root),
-        )
-        .unwrap();
-
-        assert!(child.namespace.resolve("/agent-definition").is_ok());
-        assert_eq!(
-            child
-                .host_mounts
-                .iter()
-                .find(|grant| grant.namespace_path == "/agent-definition")
-                .unwrap()
-                .host_path,
-            child_definition
-        );
-    }
-
-    #[test]
-    fn child_launch_context_passes_parent_host_mounts_only_with_explicit_handle() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("done"),
-        );
-        let parent_context = parent.namespace_environment().launch_context().unwrap();
-        let definition = temp.path().join("child-definition");
-        let spec = launch_spec(definition.clone());
-
-        let definition = host_launch_root(definition);
-        let child =
-            build_child_launch_context(parent_context, &spec, None, Some(&definition)).unwrap();
-
-        assert_eq!(child.cwd, "/mnt/source");
-        assert!(child.namespace.resolve("/mnt/source").is_ok());
-        assert!(
-            child
-                .host_mounts
-                .iter()
-                .any(|grant| grant.namespace_path == "/mnt/source")
-        );
-    }
-
-    #[test]
-    fn child_launch_rejects_cwd_inside_an_unpassed_host_mount() {
-        let temp = TempDir::new().unwrap();
-        let parent = make_parent_state(
-            &temp,
-            RecordedRequests::default(),
-            completed_response("done"),
-        );
-        let parent_context = parent.namespace_environment().launch_context().unwrap();
-        let definition = temp.path().join("child-definition");
-        let mut spec = launch_spec(definition.clone());
-        spec.handles.clear();
-
-        let definition = host_launch_root(definition);
-        let error = build_child_launch_context(
-            parent_context,
-            &spec,
-            Some("/mnt/source".to_string()),
-            Some(&definition),
-        )
-        .unwrap_err();
-        assert!(error.to_string().contains("explicit host_mounts handle"));
-    }
-
-    #[test]
-    fn child_launch_contract_normalizes_repeated_namespace_separators() {
-        let mut spec = launch_spec(PathBuf::from("/tmp/definition"));
-        spec.launch.cwd = Some(PathBuf::from("/mnt//source///docs"));
-
-        assert_eq!(
-            validate_child_launch_contract(&spec).unwrap().as_deref(),
-            Some("/mnt/source/docs")
-        );
-    }
-
-    #[tokio::test]
-    async fn child_runtime_join_keeps_running_while_activity_file_is_fresh() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("finished after file heartbeat");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let spec = launch_spec(temp.path().join("definition"));
-        let mut child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(
-                RecordingProvider::new(requests.clone(), response.clone())
-                    .with_delay(Duration::from_millis(250)),
-            ))
-        })
-        .await
-        .unwrap();
-        child.set_timeout_for_test(Duration::from_millis(200));
-        let environment = child.process_environment_for_test().clone();
-        tokio::spawn(async move {
-            for _ in 0..5 {
-                crate::runtime::ui_surfaces::heartbeat(&environment)
-                    .await
-                    .unwrap();
-                tokio::time::sleep(Duration::from_millis(35)).await;
-            }
-        });
-
-        let result = child.join().await.unwrap();
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(result.output_text, "finished after file heartbeat");
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_cancellable_aborts_pre_cancelled_launch() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("This should never run.");
-        let parent = make_parent_state(&temp, requests, response);
-        let root_dir = temp.path().join("definition");
-        let spec = launch_spec(root_dir);
-        let cancel = CancellationToken::new();
-        cancel.cancel();
-
-        let err = match spawn_child_runtime_cancellable(&parent, spec, &cancel).await {
-            Ok(_) => {
-                panic!("pre-cancelled launch should abort before returning a child controller")
-            }
-            Err(err) => err,
-        };
-
-        assert!(
-            err.to_string()
-                .contains("Child Agent Process launch cancelled")
-        );
-    }
-
-    #[test]
-    fn child_run_status_for_launch_error_maps_cancelled_launches_to_cancelled() {
-        let cancelled = anyhow::anyhow!(CHILD_AGENT_LAUNCH_CANCELLED_MESSAGE);
-        let failed = anyhow::anyhow!("Failed to submit initial child Agent Process turn");
-
-        assert_eq!(
-            child_run_status_for_launch_error(&cancelled),
-            ChildRunStatus::Cancelled
-        );
-        assert_eq!(
-            child_run_status_for_launch_error(&failed),
-            ChildRunStatus::Failed
-        );
-    }
-
-    #[tokio::test]
-    async fn child_runtime_join_returns_promptly_after_timeout() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("This should not finish before timeout.");
-        let parent = make_parent_state(&temp, requests.clone(), response.clone());
-        let root_dir = temp.path().join("definition");
-        let mut spec = launch_spec(root_dir);
-        spec.launch.timeout_secs = Some(1);
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(
-                RecordingProvider::new(requests.clone(), response.clone())
-                    .with_delay(Duration::from_secs(30)),
-            ))
-        })
-        .await
-        .unwrap();
-        let process_environment = child.process_environment_for_test().clone();
-        let process_pid = child.process_pid_for_test().to_string();
-
-        let started_at = std::time::Instant::now();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::TimedOut);
-        assert_eq!(
-            process_environment
-                .read_process_exit_code(&process_pid)
-                .await
-                .unwrap(),
-            Some(124)
-        );
-        assert!(
-            started_at.elapsed() < Duration::from_secs(8),
-            "timed-out child join should abort promptly instead of waiting for graceful shutdown"
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_resolves_package_child_agent_target() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Package child target resolved.");
-        let capability_view = capability_view_with_package_child_agent(&temp);
-        let parent = make_parent_state_with_capability_view(
-            &temp,
-            requests.clone(),
-            response.clone(),
-            capability_view,
-        );
-        let spec = SpawnSpec {
-            target: SpawnTarget::PackageChildAgent {
-                package_id: "skill:repo-review".to_string(),
-                export_name: "reviewer".to_string(),
-            },
-            launch: alan_agent_protocol::SpawnLaunchInputs {
-                task: "Review the repository changes".to_string(),
-                cwd: Some(PathBuf::from("/mnt/source")),
-                timeout_secs: Some(30),
-                ..alan_agent_protocol::SpawnLaunchInputs::default()
-            },
-            handles: vec![SpawnHandle::HostMounts],
-            runtime_overrides: alan_agent_protocol::SpawnRuntimeOverrides::default(),
-            delegated: None,
-        };
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(result.output_text, "Package child target resolved.");
-    }
-
-    #[tokio::test]
-    async fn spawn_child_runtime_resolves_package_child_agent_target_from_refreshed_view() {
-        let temp = TempDir::new().unwrap();
-        let requests = RecordedRequests::default();
-        let response = completed_response("Package child target resolved after refresh.");
-        let package_store = temp.path().join("package-store");
-        let package_root = package_store.join("repo-review");
-        std::fs::create_dir_all(&package_root).unwrap();
-        std::fs::write(
-            package_root.join("SKILL.md"),
-            r#"---
-name: Repo Review
-description: Review repository changes
----
-
-Body
-"#,
-        )
-        .unwrap();
-
-        let capability_view = crate::skills::ResolvedCapabilityView::from_package_dirs(vec![
-            crate::skills::ScopedPackageDir {
-                path: package_store,
-                scope: crate::skills::SkillScope::Descriptor,
-            },
-        ]);
-        let parent = make_parent_state_with_capability_view(
-            &temp,
-            requests.clone(),
-            response.clone(),
-            capability_view,
-        );
-
-        std::fs::create_dir_all(package_root.join("agents/reviewer")).unwrap();
-        std::fs::write(
-            package_root.join("agents/reviewer/agent.toml"),
-            "tool_repeat_limit = 4\n",
-        )
-        .unwrap();
-
-        let spec = SpawnSpec {
-            target: SpawnTarget::PackageChildAgent {
-                package_id: "skill:repo-review".to_string(),
-                export_name: "reviewer".to_string(),
-            },
-            launch: alan_agent_protocol::SpawnLaunchInputs {
-                task: "Review the repository changes".to_string(),
-                cwd: Some(PathBuf::from("/mnt/source")),
-                timeout_secs: Some(30),
-                ..alan_agent_protocol::SpawnLaunchInputs::default()
-            },
-            handles: vec![SpawnHandle::HostMounts],
-            runtime_overrides: alan_agent_protocol::SpawnRuntimeOverrides::default(),
-            delegated: None,
-        };
-
-        let child = spawn_child_runtime_with_client_factory(&parent, spec, |_| {
-            Ok(LlmClient::new(RecordingProvider::new(
-                requests.clone(),
-                response.clone(),
-            )))
-        })
-        .await
-        .unwrap();
-        let result = child.join().await.unwrap();
-
-        assert_eq!(result.status, ChildRuntimeStatus::Completed);
-        assert_eq!(
-            result.output_text,
-            "Package child target resolved after refresh."
-        );
-    }
-}
+mod tests;
